move elements to higher level
