+++ py_src/ipyelk/elements/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2021 Dane Freeman.
 # Distributed under the terms of the Modified BSD License.
 from .compound import Compound, Mark
-from .elements import Edge, Label, Node, Port, element
+from .elements import Edge, Label, Node, Port, element, ElementMetadata
 from .extended import Compartment, Partition, Record
 from .registry import Registry
 
@@ -10,6 +10,7 @@ __all__ = [
     "Compound",
     "Edge",
     "element",
+    "ElementMetadata",
     "Label",
     "Mark",
     "Node",
